pyproject.toml
[project]
name = "generate-prompts"
version = "0.1.0"
description = "Reusable LLM prompt generator & extender library"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
streamlit = ["streamlit>=1.30"]
dev = ["pytest>=8.0", "ruff>=0.4"]
